move nerf-vuer out
